User/project population script
This is an initial version of the data population script. Currently,
this script takes as input
1. The number of users to be created
2. The number of projects to be created

Currently, all the created users have projectAdmin as their assigned
roles. This will change the in the forthcoming commits.
This script is the first part of a verification test suite which can be
leveraged to check the consistency of data with various data backends.